Move SyncSettingsSnapshot to sync package; update imports
