+++ core/src/main/kotlin/at/bitfire/davdroid/sync/JtxSyncer.kt
@@ -12,7 +12,6 @@ import at.bitfire.davdroid.db.Collection
 import at.bitfire.davdroid.db.Service
 import at.bitfire.davdroid.resource.LocalJtxCollection
 import at.bitfire.davdroid.resource.LocalJtxCollectionStore
-import at.bitfire.davdroid.settings.SyncSettingsSnapshot
 import at.bitfire.synctools.storage.TaskProvider
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
