+++ core/src/main/kotlin/at/bitfire/davdroid/sync/SyncSettingsSnapshot.kt
@@ -1,21 +1,22 @@
 /*
  * Copyright © All Contributors. See LICENSE and AUTHORS in the root directory for details.
  */
-package at.bitfire.davdroid.settings
+package at.bitfire.davdroid.sync
 
 import android.accounts.Account
 import at.bitfire.davdroid.di.qualifier.IoDispatcher
+import at.bitfire.davdroid.settings.AccountSettings
 import at.bitfire.synctools.vcard.GroupMethod
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 /**
- * Immutable snapshot of the [AccountSettings] that are relevant for a sync run.
+ * Immutable snapshot of the [at.bitfire.davdroid.settings.AccountSettings] that are relevant for a sync run.
  *
- * Should be created once at the start of a sync run instead of querying [AccountSettings]
+ * Should be created once at the start of a sync run instead of querying [at.bitfire.davdroid.settings.AccountSettings]
  * again and again during the same sync. This way, settings can't change in the middle of
- * a sync run, and expensive/blocking [AccountSettings] access is reduced.
+ * a sync run, and expensive/blocking [at.bitfire.davdroid.settings.AccountSettings] access is reduced.
  */
 data class SyncSettingsSnapshot(
     // AccountSettings (same order)
